move flags to the mcserverfx package
Moving flags to the server package closer to server config parsing to
make it easier to merge configs and flags.

We have two sources of configuration: a config file and command line
flags. Commandline flags make it easier to run the server quickly for
local set ups, while the yaml configuration file is easier to use when
the server is deployed as a service.

This adds a lot of complexity when initializing the server to build the
right options structs and validating the configuration.

I'm planning to move configuration loading to be yaml driven while
having the ability to override configs using commandline flags.
