+++ mcp-server/pkg/mcpserverfx/flags.go
@@ -1,4 +1,4 @@
-package cmd
+package mcpserverfx
 
 import (
 	"errors"
